+++ test/DataFilters.UnitTests/Helpers/ExpressionsGenerators.cs
@@ -11,7 +11,9 @@ namespace DataFilters.UnitTests.Helpers
     using System.Globalization;
     using System.Linq;
 
-    public static partial class ExpressionsGenerators
+    using static GeneratorHelper;
+
+    public static class ExpressionsGenerators
     {
         public static Arbitrary<DateTimeExpression> DateTimeExpressions()
         {
@@ -115,10 +117,6 @@ namespace DataFilters.UnitTests.Helpers
                                                     : new StringValueExpression(value.Item)
                                                 )
                                                 .ToArbitrary();
-
-        private static Arbitrary<T> GetArbitraryFor<T>(Func<T, bool> filter = null) => ArbMap.Default.ArbFor<T>()
-                                                                                                     .Filter(item => filter?.Invoke(item) != false);
-
 
 
         public static Arbitrary<DurationExpression> DurationExpressions()